+++ fastlane/lib/fastlane/plugins/plugins.rb
@@ -10,4 +10,3 @@ require 'fastlane/plugins/plugin_manager'
 require 'fastlane/plugins/plugin_search'
 require 'fastlane/plugins/plugin_fetcher'
 require 'fastlane/plugins/plugin_update_manager'
-require 'fastlane/plugins/plugin_event_manager'
